# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a Project Management API",
    "version": "1.0.0",
    "title": "Project Management"
  },
  "host": "cse341-project-management.herokuapp.com",
...
  "schemes": ["https"],
  "paths": {
...
    "/teams": {
      "get": {
        "tags": ["team"],
        "summary": "Get all of the teams that the user is a part of",
        "description": "",
        "operationId": "getTeams",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Team"
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      },
      "post": {
        "tags": ["team"],
        "summary": "Create a new team",
        "description": "",
        "operationId": "addTeam",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Team object to be added",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Team"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Team successfully added"
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/teams/{teamId}": {
      "get": {
        "tags": ["team"],
        "summary": "Find team by ID",
        "description": "",
        "operationId": "getTeamById",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "description": "The team ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Team"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Team not found"
          }
        }
      },
      "put": {
        "tags": ["team"],
        "summary": "Add a user to an existing team",
        "description": "",
        "operationId": "updateTeam",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "description": "The team ID",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Email of the user to add to the team",
            "required": true,
            "schema": {
              "properties": {
                "email": {
                  "type": "string",
                  "format": "email"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Team updated successfully"
          },
          "400": {
            "description": "Invalid ID or email supplied"
          },
          "404": {
            "description": "Team not found"
          }
        }
      },
      "delete": {
        "tags": ["team"],
        "summary": "Remove yourself from a team",
        "description": "By using this endpoint, you will be removed from the specified team, but other members of the team will not be removed. If you are the last user left in the team when you leave, the team will be deleted.",
        "operationId": "removeTeam",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "description": "The id of the team you are leaving",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully removed from team"
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Team not found"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Team": {
      "type": "object",
      "required": ["name"],
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "members": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "projects": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
  }
}